Refactor account creation logic

+++ src/main/kotlin/camilyed/github/io/currencyexchangeapi/Account.kt
@@ -1,9 +1,10 @@
 package camilyed.github.io.currencyexchangeapi
 
+import java.math.BigDecimal
 import java.util.UUID
 
 data class Account(
     val id: UUID,
     val owner: String,
-    val balancePln: Double
+    val balancePln: BigDecimal
 )
