+++ src/main/kotlin/camilyed/github/io/currencyexchangeapi/CreateAccountCommand.kt
@@ -1,6 +1,8 @@
 package camilyed.github.io.currencyexchangeapi
 
+import java.math.BigDecimal
+
 data class CreateAccountCommand(
     val owner: String,
-    val initialBalance: Double
+    val initialBalance: BigDecimal
 )
